Better workaround for modules.First() issue
Apparently I'm allowed to throw an exception from the init method,
and doing so will cause LiveSplit to retry it again.
Hopefully this takes care of the instances where modules.First()
points to ntdll.dll instead of the game's main module, in a way
better than using game.MainModule (which the ASL documentation
recommends against using).